Safari-611.1.10/LayoutTests/media/video-poster-not-found-expected.txt


Test <video> element with only width attribute and the poster is missing (cannot be downloaded).

RUN(video.src = findMediaFile("video", "content/test"))
EVENT(canplaythrough)
EXPECTED (video.clientWidth == '320') OK
EXPECTED (video.clientHeight == '240') OK
END OF TEST
